ex5 edit

+++ Baekjoon/step1/ex5.java
@@ -6,8 +6,8 @@ public class ex5 {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int A = sc.nextInt();
-		int B = sc.nextInt();
+		double A = sc.nextDouble();
+		double B = sc.nextDouble();
 		
 		System.out.println(A/B);
